deduplicate shared test support helpers
